Refactor: useDisableScrollEffect
