+++ src/hooks/useDisableScrollEffect/index.css
@@ -1,0 +1,6 @@
+html.lock-scroll body {
+  height: calc(var(--window-inner-height) - 1px);
+  overflow: hidden;
+  box-sizing: border-box;
+  touch-action: none;
+}
